restoring handle for spectrumGroup drop on blank display

+++ src/python/ccpn/ui/gui/widgets/SpectrumGroupsToolBarWidget.py
@@ -29,23 +29,37 @@ __date__ = "$Date: 2017-04-07 10:28:41 +0000 (Fri, April 07, 2017) $"
 from PyQt4 import QtCore, QtGui
 from ccpn.ui.gui.widgets.Base import Base
 from ccpn.ui.gui.widgets.Button import Button
+from ccpn.ui.gui.widgets.ToolBar import ToolBar
+
+class SpectrumGroupsToolBar(ToolBar):
+  def __init__(self, parent=None, spectrumDisplay=None, spectrumGroup=None, **kwds):
+    ToolBar.__init__(self, parent=parent, **kwds)
+
+    spectrumGroupButton = SpectrumGroupsWidget(parent=parent, spectrumDisplay=spectrumDisplay, spectrumGroup=spectrumGroup)
 
 
-class SpectrumGroupsToolBar(QtGui.QToolBar, Base):
-  def __init__(self, parent=None, project=None, strip=None, spectrumGroupPid=None, **kw):
-    QtGui.QWidget.__init__(self, parent)
+class ToolButton(QtGui.QToolButton, Base):
+
+  def __init__(self, parent=None, text='', **kw):
+
+
+    QtGui.QToolButton.__init__(self, parent)
     Base.__init__(self, **kw)
 
-    spectrumGroupButton = SpectrumGroupsWidget(self, project=project, strip=strip, spectrumGroupPid=spectrumGroupPid)
-    self.addWidget(spectrumGroupButton)
+    self.setText(text)
+
 
 class SpectrumGroupsWidget(QtGui.QWidget):
-  def __init__(self, parent=None, project=None, strip=None, spectrumGroupPid=None, **kw):
+  def __init__(self, parent=None, spectrumDisplay=None, spectrumGroup=None, **kw):
     QtGui.QWidget.__init__(self, parent)
-    self.project = project
-    self.strip = strip
-    self.spectrumGroup = self.project.getByPid(spectrumGroupPid)
+
+    self.spectrumDisplay = spectrumDisplay
+    if len(spectrumDisplay.strips) > 0:
+      self.strip = spectrumDisplay.strips[0]
+
+    self.spectrumGroup = spectrumGroup
     self.spectrumGroupButton = Button(self, text=self.spectrumGroup.id,toggle=True)
+
     self.spectrumGroupButton.setChecked(True)
     self.spectrumGroupButton.setMinimumSize(40,33)
     self.spectrumGroupButton.toggled.connect(self.toggleSpectrumGroups)
@@ -60,8 +74,8 @@ class SpectrumGroupsWidget(QtGui.QWidget):
     self.popMenu.addAction(self.peakListCheckBox)
 
     self.spectrumGroupPeakLists = [spectrum.peakLists[0] for spectrum in self.spectrumGroup.spectra]
-    self.peakListViews = [peakListView for peakListView in self.strip.spectrumDisplay.peakListViews ]
-    self.peakListViewDisplayed = [peakListView.peakList for peakListView in self.strip.spectrumDisplay.peakListViews ]
+    self.peakListViews = [peakListView for peakListView in self.spectrumDisplay.peakListViews ]
+    self.peakListViewDisplayed = [peakListView.peakList for peakListView in self.spectrumDisplay.peakListViews ]
 
   def onContextMenu(self, points):
     positions = self.spectrumGroupButton.mapToGlobal(points)
@@ -69,28 +83,30 @@ class SpectrumGroupsWidget(QtGui.QWidget):
     self.popMenu.exec()
 
   def toggleSpectrumGroups(self):
-    spectrumViews = [spectrumView for spectrumView in self.strip.spectrumViews
-                     if spectrumView.spectrum in self.spectrumGroup.spectra]
+    if self.strip is not None:
+      spectrumViews = [spectrumView for spectrumView in self.strip.spectrumViews
+                       if spectrumView.spectrum in self.spectrumGroup.spectra]
 
-    if self.spectrumGroupButton.isChecked():
-      for spectrumView in spectrumViews:
-        spectrumView.setVisible(True)
-        if hasattr(spectrumView, 'plot'):
-          spectrumView.plot.show()
-      self.showPeakList()
+      if self.spectrumGroupButton.isChecked():
+        for spectrumView in spectrumViews:
+          spectrumView.setVisible(True)
+          if hasattr(spectrumView, 'plot'):
+            spectrumView.plot.show()
+        self.showPeakList()
 
-    else:
-      for spectrumView in spectrumViews:
-        spectrumView.setVisible(False)
-        if hasattr(spectrumView, 'plot'):
-          spectrumView.plot.hide()
-      self.hidePeakLists()
+      else:
+        for spectrumView in spectrumViews:
+          spectrumView.setVisible(False)
+          if hasattr(spectrumView, 'plot'):
+            spectrumView.plot.hide()
+        self.hidePeakLists()
 
   def deleteSpectrumGroup(self):
-    self.spectrumGroupButton.deleteLater()
-    for spectrumView in self.strip.spectrumViews:
-      if spectrumView.spectrum in self.spectrumGroup.spectra:
-        spectrumView.delete()
+    if self.strip is not None:
+      self.spectrumGroupButton.deleteLater()
+      for spectrumView in self.strip.spectrumViews:
+        if spectrumView.spectrum in self.spectrumGroup.spectra:
+          spectrumView.delete()
 
   def showHidePeakListView(self):
     if self.peakListCheckBox.isChecked():
@@ -100,12 +116,14 @@ class SpectrumGroupsWidget(QtGui.QWidget):
 
   def hidePeakLists(self):
     for peakList in self.spectrumGroupPeakLists:
-      for peakListView in self.strip.spectrumDisplay.peakListViews:
-        if peakList == peakListView.peakList:
-          peakListView.setVisible(False)
+      if self.spectrumDisplay is not None:
+        for peakListView in self.spectrumDisplay.peakListViews:
+          if peakList == peakListView.peakList:
+            peakListView.setVisible(False)
 
   def showPeakList(self):
     for peakList in self.spectrumGroupPeakLists:
-      for peakListView in self.strip.spectrumDisplay.peakListViews:
-        if peakList == peakListView.peakList:
-          peakListView.setVisible(True)
+      if self.spectrumDisplay is not None:
+        for peakListView in self.spectrumDisplay.peakListViews:
+          if peakList == peakListView.peakList:
+            peakListView.setVisible(True)
